requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
crispy-bootstrap4==2024.1
crispy-bootstrap5==2025.4
Django>=5.0.2
django-crispy-forms>=2.1
frozenlist==1.6.0
idna==3.10
multidict==6.4.4
Pillow>=10.2.0
propcache==0.3.1
PyJWT==2.10.1
pypng==0.20220715.0
python-dotenv>=1.0.0
qrcode>=7.4.2
requests==2.32.3
sqlparse==0.5.3
twilio==8.12.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
